+++ NEWS.md
@@ -1,4 +1,4 @@
-# targetd 0.8
+# targeted 0.8
 
 Development version
 
@@ -7,6 +7,8 @@ Development version
 - feature(learner_cox): mets::phreg wrapper.
 - disabling openMP in RcppArmadillo
 - bugfix: issue with `cate` with repeated cross-fitting fixed.
+- `cate`: new argument `effect.scale` that allows for one-step estimation
+  directly treatment effects on the log og log-odds scale.
 
 # targeted 0.7.1